adding Parallel and CUDA unit tests for mumps and slu
