+++ tests/unit/CMakeLists.txt
@@ -22,6 +22,7 @@ set(UNIT_TESTS_SRCS
   general/test_text.cpp
   general/test_zlib.cpp
   linalg/test_complex_operator.cpp
+  linalg/test_direct_solvers.cpp
   linalg/test_hypre_ilu.cpp
   linalg/test_ilu.cpp
   linalg/test_matrix_block.cpp
@@ -34,7 +35,6 @@ set(UNIT_TESTS_SRCS
   linalg/test_ode2.cpp
   linalg/test_operator.cpp
   linalg/test_cg_indefinite.cpp
-  linalg/test_umf.cpp
   linalg/test_vector.cpp
   mesh/test_mesh.cpp
   mesh/test_ncmesh.cpp
